backend/requirements.txt
# Core MCP server
fastmcp==3.2.4
mcp==1.27.0

# Web framework (frontend API)
Flask==3.1.3
flask-cors==6.0.2

# Desktop UI
pywebview==6.2

# Screen capture (Eye MCP)
Pillow>=10.0

# Computer control (Computer MCP)
pyautogui>=0.9.54
pyperclip>=1.11.0
PyGetWindow>=0.0.9

# Vector database client
qdrant-client==1.17.1

# Memory engine (mem0)
mem0ai>=2.0.0

# Embedding model
sentence-transformers==5.4.1
transformers==5.5.4
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.10.2

# PyTorch - CPU 版本（默认）
# 如需 GPU，自行安装：pip install torch --index-url https://download.pytorch.org/whl/cu124
torch

# Settings & validation
pydantic==2.13.0
pydantic-settings==2.13.1

# System monitoring
psutil==7.2.2
nvidia-ml-py==12.575.51
scikit-learn==1.8.0
scipy==1.17.1
tqdm==4.67.3
python-dotenv==1.2.2
